% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{beta.post}
\alias{beta.post}
\title{beta.post}
\usage{
beta.post(beta.p, p, beta, Sigma.b, Y, C, X, phi, lambda)
}
\arguments{
\item{beta.p}{current value of the pth element of beta}

\item{p}{pth element}

\item{beta}{current value of beta}

\item{Sigma.b}{variance estimate of beta}

\item{Y}{response variable}

\item{C}{censoring indicator}

\item{X}{covariates for beta}

\item{phi}{current value of phi}

\item{lambda}{current value of lambda}
}
\value{
log- posterior density of beta
}
\description{
log-posterior distribution of beta with pth element fixed as beta.p
}
