% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{beta.slice.sampling}
\alias{beta.slice.sampling}
\title{beta.slice.sampling}
\usage{
beta.slice.sampling(beta, Sigma.b, Y, C, X, phi, lambda, w, m = 100)
}
\arguments{
\item{beta}{previous value of beta used in the slice sampling}

\item{Sigma.b}{variance estimate of beta}

\item{Y}{response variable}

\item{C}{censoring indicator}

\item{X}{covariates for beta}

\item{phi}{current value of phi}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for beta
}
