% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{betas.post}
\alias{betas.post}
\title{betas.post}
\usage{
betas.post(betas.p, p, Sigma.b, Y, X, betas, Z, gammas, C, lambda)
}
\arguments{
\item{betas.p}{current value of the pth element of betas}

\item{p}{pth element}

\item{Sigma.b}{variance estimate of betas}

\item{Y}{response variable}

\item{X}{covariates for betas}

\item{betas}{current value of betas}

\item{Z}{covariates for gammas}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of betas with pth element fixed as betas.p
}
