% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{betas.slice.sampling}
\alias{betas.slice.sampling}
\title{betas.slice.sampling}
\usage{
betas.slice.sampling(Sigma.b, Y, X, betas, Z, gammas, C, lambda, w, m)
}
\arguments{
\item{Sigma.b}{variance estimate of betas}

\item{Y}{response variable}

\item{X}{covariates for betas}

\item{betas}{current value of betas}

\item{Z}{covariates for gammas}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for betas
}
