% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{gammas.post}
\alias{gammas.post}
\title{gammas.post}
\usage{
gammas.post(gammas.p, p, gammass, Sigma.g, Y, C, XB, Z, lambda)
}
\arguments{
\item{gammas.p}{current value of the pth element of gammas}

\item{p}{pth element}

\item{gammass}{current value of gammas}

\item{Sigma.g}{variance estimate of gammas}

\item{Y}{response variable}

\item{C}{censoring indicator}

\item{XB}{current value of XB}

\item{Z}{covariates for gammas}

\item{lambda}{current value of lambda}
}
\value{
log- posterior density of beta
}
\description{
log-posterior distribution of gammas with pth element fixed as gammas.p
}
