% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{lambda.post}
\alias{lambda.post}
\title{lambda.post}
\usage{
lambda.post(lambda, Y, C, XB, phi, a = 1, b = 1)
}
\arguments{
\item{lambda}{current value of lambda}

\item{Y}{response variable}

\item{C}{censoring indicator}

\item{XB}{current value of XB}

\item{phi}{current value of phi}

\item{a}{shape parameter of gammas prior}

\item{b}{scale parameter of gammas prior}
}
\value{
log- posterior density of beta
}
\description{
log-posterior distribution of lambda
}
