% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{lambda.slice.sampling}
\alias{lambda.slice.sampling}
\title{lambda.slice.sampling}
\usage{
lambda.slice.sampling(Y, X, betas, Z, gammas, C, lambda, w, m, lower = 0 +
  10^(-10), upper = +Inf)
}
\arguments{
\item{Y}{response variable}

\item{X}{covariates for betas}

\item{betas}{current value of betas}

\item{Z}{covariates for gammas}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for lambda
}
