% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{univ.beta.slice.sampling}
\alias{univ.beta.slice.sampling}
\title{univ.beta.slice.sampling}
\usage{
univ.beta.slice.sampling(beta.p, p, beta, Sigma.b, Y, C, X, phi, lambda, w,
  m = 100, lower = -Inf, upper = +Inf)
}
\arguments{
\item{beta.p}{current value of the pth element of beta}

\item{p}{pth element}

\item{beta}{current value of beta}

\item{Sigma.b}{variance estimate of beta}

\item{Y}{response variable}

\item{C}{censoring indicator}

\item{X}{covariates for beta}

\item{phi}{current value of phi}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for beta.p
}
