% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{univ.gammas.slice.sampling}
\alias{univ.gammas.slice.sampling}
\title{univ.gammas.slice.sampling}
\usage{
univ.gammas.slice.sampling(gammas.p, p, gammass, Sigma.g, Y, C, XB, Z, lambda,
  w, m = 100, lower = -Inf, upper = +Inf)
}
\arguments{
\item{gammas.p}{current value of the pth element of gammas}

\item{p}{pth element}

\item{gammass}{current value of gammas}

\item{Sigma.g}{variance estimate of gammas}

\item{Y}{response variable}

\item{C}{censoring indicator}

\item{XB}{current value of XB}

\item{Z}{covariates for gammas}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for gammas.p
}
