% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{gammas.slice.sampling}
\alias{gammas.slice.sampling}
\title{gammass.slice.sampling}
\usage{
gammas.slice.sampling(gammass, Sigma.g, Y, C, XB, Z, lambda, w, m = 100)
}
\arguments{
\item{gammass}{previous value of gammass used in the slice sampling}

\item{Sigma.g}{variance estimate of gammass}

\item{Y}{response variable}

\item{C}{censoring indicator}

\item{XB}{current value of XB}

\item{Z}{covariates for gammas}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for gammass
}
