% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{lambda.post}
\alias{lambda.post}
\title{lambda.post}
\usage{
lambda.post(Y, X, betas, Z, gammas, C, lambda, a = 1, b = 1)
}
\arguments{
\item{Y}{response variable}

\item{X}{covariates for betas}

\item{betas}{current value of betas}

\item{Z}{covariates for gammas}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{a}{shape parameter of gammas prior}

\item{b}{scale parameter of gammas prior}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of lambda
}
