% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesOFsurv.R
\name{univ.gammas.slice.sampling}
\alias{univ.gammas.slice.sampling}
\title{univ.gammas.slice.sampling}
\usage{
univ.gammas.slice.sampling(gammas.p, p, Sigma.g, Y, X, betas, Z, gammas, C,
  lambda, w, m, lower = -Inf, upper = +Inf)
}
\arguments{
\item{gammas.p}{current value of the pth element of gammas}

\item{p}{pth element}

\item{Sigma.g}{variance estimate of gammas}

\item{Y}{response variable}

\item{X}{covariates for betas}

\item{betas}{current value of betas}

\item{Z}{covariates for gammas}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for gammas.p
}
